requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
pydantic>=2.9.0
pydantic-settings>=2.0.0
APScheduler>=3.10.0
pytz>=2024.1
sentry-sdk[fastapi]>=2.15.0
httpx>=0.27.0
python-dotenv>=1.0.0
opencv-python-headless>=4.10.0
pytesseract>=0.3.10
Pillow>=10.4.0
openai>=1.40.0
stripe>=10.0.0
pyyaml>=6.0.0
pyjwt
passlib>=1.7.4
python-multipart>=0.0.9
feedparser>=6.0.11
pyotp>=2.9.0
email-validator>=2.2.0
redis>=5.0.0
requests>=2.32.0
joblib>=1.4.2
scikit-learn>=1.5.2
cryptography>=42.0.0
pinecone-client>=3.0.0
prometheus-client>=0.21.0
